deployment/aws/requirements.txt
# Lightweight requirements for Vercel deployment
Flask==3.0.0
python-telegram-bot==22.7
python-dotenv==1.2.1
requests==2.32.5
pandas==2.2.0
numpy==1.26.4
